requirements.txt
alembic==0.8.3
aniso8601==1.0.0
arrow==0.7.0
Babel==2.1.1
bcrypt==2.0.0
cffi==1.2.1
docutils==0.12
fake-factory==0.5.2
Flask==0.10.1
Flask-Classy==0.6.10
flask-marshmallow==0.6.2
Flask-Migrate==1.6.0
Flask-RESTful==0.3.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
LinkHeader==0.4.3
Mako==1.0.2
MarkupSafe==0.23
marshmallow==2.1.3
marshmallow-sqlalchemy==0.6.0
mixer==5.2.2
passlib==1.6.5
pycparser==2.14
Pygments==2.0.2
python-dateutil==2.4.2
python-editor==0.4
pytz==2015.6
PyYAML==3.11
requests==2.8.1
six==1.10.0
snowballstemmer==1.2.0
speaklater==1.3
SQLAlchemy==1.0.8
SQLAlchemy-Utils==0.31.0
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
